+++ tests/tests.yaml
@@ -128,6 +128,10 @@ answer_tests:
   local_axialpix_006:
     - yt/geometry/coordinates/tests/test_axial_pixelization.py:test_axial_pixelization
 
+  local_cylindrical_background_000:
+    - yt/geometry/coordinates/tests/test_cylindrical_coordinates.py:test_noise_plot_lin
+    - yt/geometry/coordinates/tests/test_cylindrical_coordinates.py:test_noise_plot_log
+
   local_particle_trajectory_001:
     - yt/data_objects/tests/test_particle_trajectories.py
